Separate plan and plan_v2 scan modules
