+++ vortex-layout/src/scan/mod.rs
@@ -5,12 +5,12 @@ pub mod arrow;
 mod filter;
 pub mod layout;
 pub mod multi;
-pub mod plan;
+mod plan;
+pub mod plan_v2;
 pub mod repeated_scan;
 pub mod scan_builder;
 pub mod split_by;
 mod splits;
-mod tasks;
 #[cfg(test)]
 mod test;
