+++ vortex-layout/src/scan/plan.rs
@@ -1,71 +1,196 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-//! Heap-allocated physical scan planning over the established layout-reader execution API.
-//!
-//! # Design direction
-//!
-//! Planning should begin with a heap-allocated source plan produced by a layout. Applying an
-//! expression to that source returns another plan, and optimizing that derived plan returns another
-//! plan. Execution receives only a row range and mask; expressions do not cross the
-//! planning/execution boundary.
-//!
-//! A scan retains its existing N+1 decomposition: one independently pushed and optimized plan for
-//! each filter conjunct, plus one independently pushed and optimized projection plan. These plans
-//! must remain separate so their reads can be scheduled concurrently. They are not combined into a
-//! single filter-then-projection operator tree.
-//!
-//! The current [`LayoutReaderScanPlan`](crate::LayoutReaderScanPlan) is a compatibility source plan:
-//! it supports the plan-to-plan API but delegates execution to the existing [`crate::LayoutReader`].
-//! Layout-specific source plans can replace it incrementally.
+//! Split scanning task implementation.
 
+use std::ops::BitAnd;
+use std::sync::Arc;
+
+use bit_vec::BitVec;
+use futures::FutureExt;
+use futures::future::BoxFuture;
+use vortex_array::ArrayRef;
+use vortex_array::MaskFuture;
+use vortex_array::expr::Expression;
 use vortex_error::VortexResult;
-use vortex_error::vortex_bail;
+use vortex_mask::Mask;
+use vortex_scan::row_mask::RowMask;
 
-/// Environment variable selecting the scan planning implementation.
-pub const SCAN_IMPL_ENV: &str = "VORTEX_SCAN_IMPL";
+use crate::LayoutReaderRef;
+use crate::scan::filter::FilterExpr;
 
-/// Returns whether heap-allocated planning is enabled for this process.
+pub type TaskFuture<A> = BoxFuture<'static, VortexResult<A>>;
+
+pub(crate) struct Plan {
+    layout_reader: LayoutReaderRef,
+    projection: Expression,
+    filter: Option<Expression>,
+}
+
+impl Plan {
+    pub(crate) fn new(
+        layout_reader: LayoutReaderRef,
+        projection: Expression,
+        filter: Option<Expression>,
+    ) -> Self {
+        Self {
+            layout_reader,
+            projection,
+            filter,
+        }
+    }
+
+    pub(crate) fn task_context<A>(
+        &self,
+        mapper: Arc<dyn Fn(ArrayRef) -> VortexResult<A> + Send + Sync>,
+    ) -> Arc<TaskContext<A>> {
+        Arc::new(TaskContext {
+            filter: self
+                .filter
+                .clone()
+                .map(|filter| Arc::new(FilterExpr::new(filter))),
+            reader: Arc::clone(&self.layout_reader),
+            projection: self.projection.clone(),
+            mapper,
+        })
+    }
+}
+
+/// Logic for executing a single split reading task.
+/// N.B. read_mask should be evaluated against all_false() before calling this
+/// method to avoid creating an empty TaskFuture.
 ///
-/// The existing V1 path remains the default on this extraction branch. Set
-/// `VORTEX_SCAN_IMPL=v2` to exercise the planning path with the same execution implementation.
-pub fn planned_scan_enabled() -> VortexResult<bool> {
-    match std::env::var(SCAN_IMPL_ENV) {
-        Ok(value) => parse_scan_impl(&value),
-        Err(std::env::VarError::NotPresent) => Ok(false),
-        Err(std::env::VarError::NotUnicode(value)) => {
-            vortex_bail!("{SCAN_IMPL_ENV} must be valid unicode, got {value:?}")
+/// # Task execution flow
+///
+/// First, the task's row range (split) is intersected with the global file row-range requested,
+/// if any.
+///
+/// The intersected row range is then further reduced via expression-based pruning. After pruning
+/// has eliminated more blocks, the full filter is executed over the remainder of the split.
+///
+/// This mask is then provided to the reader to perform a filtered projection over the split data,
+/// finally mapping the Vortex columnar record batches into some result type `A`.
+pub fn split_exec<A: 'static + Send>(
+    ctx: Arc<TaskContext<A>>,
+    read_mask: RowMask,
+    limit: Option<&mut u64>,
+) -> VortexResult<TaskFuture<Option<A>>> {
+    let row_range = read_mask.row_range();
+    let row_mask = read_mask.mask().clone();
+
+    let filter_mask = match ctx.filter.as_ref() {
+        // No filter == immediate mask
+        None => {
+            let row_mask = match limit {
+                Some(l) if *l == 0 => Mask::new_false(row_mask.len()),
+                Some(l) => {
+                    let true_count = row_mask.true_count();
+                    let mask_limit = usize::try_from(*l)
+                        .map(|l| l.min(true_count))
+                        .unwrap_or(true_count);
+                    let row_mask = row_mask.limit(mask_limit);
+                    *l -= mask_limit as u64;
+                    row_mask
+                }
+                None => row_mask,
+            };
+
+            MaskFuture::ready(row_mask)
         }
-    }
+        Some(filter) => {
+            // NOTE: it's very important that the pruning and filter evaluations are built OUTSIDE
+            // the future. Registering these row ranges eagerly is a hint to the IO system that
+            // we want to start prefetching the IO for this split.
+            let reader = Arc::clone(&ctx.reader);
+            let filter = Arc::clone(filter);
+            let row_range = row_range.clone();
+
+            MaskFuture::new(row_mask.len(), async move {
+                let mut mask = row_mask;
+                let mut dynamic_versions = vec![None; filter.conjuncts().len()];
+
+                // TODO(ngates): we could use FuturedUnordered to intersect the masks in parallel.
+                for (idx, conjunct) in filter.conjuncts().iter().enumerate() {
+                    if mask.all_false() {
+                        return Ok(mask);
+                    }
+
+                    // Store the latest version of the dynamic expression prior to pruning.
+                    // We will re-run the pruning later if the version has changed in the meantime.
+                    dynamic_versions[idx] = filter.dynamic_updates(idx).map(|du| du.version());
+
+                    let conjunct_mask = reader
+                        .pruning_evaluation(&row_range, conjunct, mask.clone())?
+                        .await?;
+                    mask = mask.bitand(&conjunct_mask);
+                }
+
+                // Now we loop through the conjuncts in the preferred order and evaluate them.
+                let mut remaining = BitVec::from_elem(filter.conjuncts().len(), true);
+                while let Some(idx) = filter.next_conjunct(&remaining) {
+                    remaining.set(idx, false);
+                    if mask.all_false() {
+                        return Ok(mask);
+                    }
+
+                    let conjunct = &filter.conjuncts()[idx];
+
+                    // If the dynamic expression has changed since pruning, re-run the pruning.
+                    // Store the dynamic update once to avoid TOCTOU race condition
+                    let current_version = filter.dynamic_updates(idx).map(|du| du.version());
+                    if let Some(dv) = current_version
+                        && dynamic_versions[idx].is_none_or(|v| v < dv)
+                    {
+                        // The dynamic expression has been updated, re-run the pruning.
+                        dynamic_versions[idx] = Some(dv);
+                        let conjunct_mask = reader
+                            .pruning_evaluation(&row_range, conjunct, mask.clone())?
+                            .await?;
+                        mask = mask.bitand(&conjunct_mask);
+                    }
+                    if mask.all_false() {
+                        return Ok(mask);
+                    }
+
+                    let conjunct_mask = reader
+                        .filter_evaluation(&row_range, conjunct, MaskFuture::ready(mask))?
+                        .await?;
+                    filter.report_selectivity(idx, conjunct_mask.density());
+
+                    // Filter evaluations return a mask already intersected with the input mask.
+                    mask = conjunct_mask;
+                }
+
+                Ok(mask)
+            })
+        }
+    };
+
+    // Step 4: execute the projection, only at the mask for rows which match the filter
+    let projection_future =
+        ctx.reader
+            .projection_evaluation(&row_range, &ctx.projection, filter_mask.clone())?;
+
+    let mapper = Arc::clone(&ctx.mapper);
+    let array_fut = async move {
+        let mask = filter_mask.await?;
+        if mask.all_false() {
+            return Ok(None);
+        }
+
+        let array = projection_future.await?;
+        mapper(array).map(Some)
+    };
+
+    Ok(array_fut.boxed())
 }
 
-fn parse_scan_impl(value: &str) -> VortexResult<bool> {
-    match value {
-        "" | "v1" | "legacy" | "layout-reader" => Ok(false),
-        "v2" | "planned" | "scan-plan" => Ok(true),
-        other => vortex_bail!(
-            "{SCAN_IMPL_ENV} must be one of v1, legacy, layout-reader, v2, planned, or scan-plan, got {other:?}"
-        ),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn scan_impl_accepts_v1_and_v2_values() -> VortexResult<()> {
-        for value in ["", "v1", "legacy", "layout-reader"] {
-            assert!(!parse_scan_impl(value)?);
-        }
-        for value in ["v2", "planned", "scan-plan"] {
-            assert!(parse_scan_impl(value)?);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn scan_impl_rejects_unknown_value() {
-        assert!(parse_scan_impl("unknown").is_err());
-    }
+/// Information needed to execute a single split task.
+///
+/// Row selection is evaluated before creating a split task so it's not included
+pub(crate) struct TaskContext<A> {
+    filter: Option<Arc<FilterExpr>>,
+    reader: LayoutReaderRef,
+    projection: Expression,
+    mapper: Arc<dyn Fn(ArrayRef) -> VortexResult<A> + Send + Sync>,
 }
