--- vortex-layout/src/scan/tasks.rs
@@ -1,288 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-//! Split scanning task implementation.
-
-use std::ops::BitAnd;
-use std::sync::Arc;
-
-use bit_vec::BitVec;
-use futures::FutureExt;
-use futures::future::BoxFuture;
-use vortex_array::ArrayRef;
-use vortex_array::MaskFuture;
-use vortex_array::expr::Expression;
-use vortex_error::VortexResult;
-use vortex_mask::Mask;
-use vortex_scan::row_mask::RowMask;
-
-use crate::LayoutReader;
-use crate::ScanPlanRef;
-use crate::scan::filter::FilterExpr;
-
-pub type TaskFuture<A> = BoxFuture<'static, VortexResult<A>>;
-
-/// Logic for executing a single split reading task.
-/// N.B. read_mask should be evaluated against all_false() before calling this
-/// method to avoid creating an empty TaskFuture.
-///
-/// # Task execution flow
-///
-/// First, the task's row range (split) is intersected with the global file row-range requested,
-/// if any.
-///
-/// The intersected row range is then further reduced via expression-based pruning. After pruning
-/// has eliminated more blocks, the full filter is executed over the remainder of the split.
-///
-/// This mask is then provided to the reader to perform a filtered projection over the split data,
-/// finally mapping the Vortex columnar record batches into some result type `A`.
-pub fn split_exec<A: 'static + Send>(
-    ctx: Arc<TaskContext<A>>,
-    read_mask: RowMask,
-    limit: Option<&mut u64>,
-) -> VortexResult<TaskFuture<Option<A>>> {
-    let row_range = read_mask.row_range();
-    let row_mask = read_mask.mask().clone();
-
-    let filter_mask = match ctx.filter.as_ref() {
-        // No filter == immediate mask
-        None => {
-            let row_mask = match limit {
-                Some(l) if *l == 0 => Mask::new_false(row_mask.len()),
-                Some(l) => {
-                    let true_count = row_mask.true_count();
-                    let mask_limit = usize::try_from(*l)
-                        .map(|l| l.min(true_count))
-                        .unwrap_or(true_count);
-                    let row_mask = row_mask.limit(mask_limit);
-                    *l -= mask_limit as u64;
-                    row_mask
-                }
-                None => row_mask,
-            };
-
-            MaskFuture::ready(row_mask)
-        }
-        Some(filter) => {
-            // NOTE: it's very important that the pruning and filter evaluations are built OUTSIDE
-            // the future. Registering these row ranges eagerly is a hint to the IO system that
-            // we want to start prefetching the IO for this split.
-            let reader = Arc::clone(&ctx.reader);
-            let filter = Arc::clone(filter);
-            let row_range = row_range.clone();
-
-            MaskFuture::new(row_mask.len(), async move {
-                let mut mask = row_mask;
-                let mut dynamic_versions = vec![None; filter.conjuncts().len()];
-
-                // TODO(ngates): we could use FuturedUnordered to intersect the masks in parallel.
-                for (idx, conjunct) in filter.conjuncts().iter().enumerate() {
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    // Store the latest version of the dynamic expression prior to pruning.
-                    // We will re-run the pruning later if the version has changed in the meantime.
-                    dynamic_versions[idx] = filter.dynamic_updates(idx).map(|du| du.version());
-
-                    let conjunct_mask = reader
-                        .pruning_evaluation(&row_range, conjunct, mask.clone())?
-                        .await?;
-                    mask = mask.bitand(&conjunct_mask);
-                }
-
-                // Now we loop through the conjuncts in the preferred order and evaluate them.
-                let mut remaining = BitVec::from_elem(filter.conjuncts().len(), true);
-                while let Some(idx) = filter.next_conjunct(&remaining) {
-                    remaining.set(idx, false);
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    let conjunct = &filter.conjuncts()[idx];
-
-                    // If the dynamic expression has changed since pruning, re-run the pruning.
-                    // Store the dynamic update once to avoid TOCTOU race condition
-                    let current_version = filter.dynamic_updates(idx).map(|du| du.version());
-                    if let Some(dv) = current_version
-                        && dynamic_versions[idx].is_none_or(|v| v < dv)
-                    {
-                        // The dynamic expression has been updated, re-run the pruning.
-                        dynamic_versions[idx] = Some(dv);
-                        let conjunct_mask = reader
-                            .pruning_evaluation(&row_range, conjunct, mask.clone())?
-                            .await?;
-                        mask = mask.bitand(&conjunct_mask);
-                    }
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    let conjunct_mask = reader
-                        .filter_evaluation(&row_range, conjunct, MaskFuture::ready(mask))?
-                        .await?;
-                    filter.report_selectivity(idx, conjunct_mask.density());
-
-                    // Filter evaluations return a mask already intersected with the input mask.
-                    mask = conjunct_mask;
-                }
-
-                Ok(mask)
-            })
-        }
-    };
-
-    // Step 4: execute the projection, only at the mask for rows which match the filter
-    let projection_future =
-        ctx.reader
-            .projection_evaluation(&row_range, &ctx.projection, filter_mask.clone())?;
-
-    let mapper = Arc::clone(&ctx.mapper);
-    let array_fut = async move {
-        let mask = filter_mask.await?;
-        if mask.all_false() {
-            return Ok(None);
-        }
-
-        let array = projection_future.await?;
-        mapper(array).map(Some)
-    };
-
-    Ok(array_fut.boxed())
-}
-
-/// Execute one split using expression-bound physical scan plans.
-///
-/// The execution order intentionally mirrors [`split_exec`]. Expressions were consumed during
-/// planning, so execution selects a predicate or projection plan without passing an expression.
-pub fn planned_split_exec<A: 'static + Send>(
-    ctx: Arc<PlannedTaskContext<A>>,
-    read_mask: RowMask,
-    limit: Option<&mut u64>,
-) -> VortexResult<TaskFuture<Option<A>>> {
-    let row_range = read_mask.row_range();
-    let row_mask = read_mask.mask().clone();
-
-    let filter_mask = match ctx.filter.as_ref() {
-        None => {
-            let row_mask = match limit {
-                Some(l) if *l == 0 => Mask::new_false(row_mask.len()),
-                Some(l) => {
-                    let true_count = row_mask.true_count();
-                    let mask_limit = usize::try_from(*l)
-                        .map(|l| l.min(true_count))
-                        .unwrap_or(true_count);
-                    let row_mask = row_mask.limit(mask_limit);
-                    *l -= mask_limit as u64;
-                    row_mask
-                }
-                None => row_mask,
-            };
-
-            MaskFuture::ready(row_mask)
-        }
-        Some(filter) => {
-            if filter.conjuncts().len() != ctx.predicates.len() {
-                vortex_error::vortex_bail!(
-                    "physical predicate count {} does not match conjunct count {}",
-                    ctx.predicates.len(),
-                    filter.conjuncts().len()
-                );
-            }
-
-            let ctx = Arc::clone(&ctx);
-            let filter = Arc::clone(filter);
-            let row_range = row_range.clone();
-
-            MaskFuture::new(row_mask.len(), async move {
-                let mut mask = row_mask;
-                let mut dynamic_versions = vec![None; filter.conjuncts().len()];
-
-                for (idx, predicate) in ctx.predicates.iter().enumerate() {
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    dynamic_versions[idx] = filter.dynamic_updates(idx).map(|du| du.version());
-                    let conjunct_mask = predicate
-                        .pruning_evaluation(&row_range, mask.clone())?
-                        .await?;
-                    mask = mask.bitand(&conjunct_mask);
-                }
-
-                let mut remaining = BitVec::from_elem(filter.conjuncts().len(), true);
-                while let Some(idx) = filter.next_conjunct(&remaining) {
-                    remaining.set(idx, false);
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    let current_version = filter.dynamic_updates(idx).map(|du| du.version());
-                    if let Some(version) = current_version
-                        && dynamic_versions[idx].is_none_or(|old| old < version)
-                    {
-                        dynamic_versions[idx] = Some(version);
-                        let conjunct_mask = ctx.predicates[idx]
-                            .pruning_evaluation(&row_range, mask.clone())?
-                            .await?;
-                        mask = mask.bitand(&conjunct_mask);
-                    }
-                    if mask.all_false() {
-                        return Ok(mask);
-                    }
-
-                    let conjunct_mask = ctx.predicates[idx]
-                        .filter_evaluation(&row_range, MaskFuture::ready(mask))?
-                        .await?;
-                    filter.report_selectivity(idx, conjunct_mask.density());
-                    mask = conjunct_mask;
-                }
-
-                Ok(mask)
-            })
-        }
-    };
-
-    let projection_future = ctx
-        .projection
-        .projection_evaluation(&row_range, filter_mask.clone())?;
-
-    let mapper = Arc::clone(&ctx.mapper);
-    let array_fut = async move {
-        let mask = filter_mask.await?;
-        if mask.all_false() {
-            return Ok(None);
-        }
-
-        let array = projection_future.await?;
-        mapper(array).map(Some)
-    };
-
-    Ok(array_fut.boxed())
-}
-
-/// Information needed to execute a single split task.
-///
-/// Row selection is evaluated before creating a split task so it's not included
-pub struct TaskContext<A> {
-    /// The shared filter expression.
-    pub filter: Option<Arc<FilterExpr>>,
-    /// The layout reader.
-    pub reader: Arc<dyn LayoutReader>,
-    /// The projection expression to apply to gather the scanned rows.
-    pub projection: Expression,
-    /// Function that maps into an A.
-    pub mapper: Arc<dyn Fn(ArrayRef) -> VortexResult<A> + Send + Sync>,
-}
-
-/// Information needed to execute one split from a prepared physical scan plan.
-pub struct PlannedTaskContext<A> {
-    /// V1 dynamic-filter tracking and adaptive conjunct ordering.
-    pub filter: Option<Arc<FilterExpr>>,
-    /// One physical plan per filter conjunct.
-    pub predicates: Vec<ScanPlanRef>,
-    /// Physical projection plan.
-    pub projection: ScanPlanRef,
-    /// Function that maps the projected array into an output.
-    pub mapper: Arc<dyn Fn(ArrayRef) -> VortexResult<A> + Send + Sync>,
-}
